requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
bottle==0.13.3
certifi==2025.4.26
#cffi==1.17.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
bcrypt>=4.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1

idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
#jiter==0.9.0
MarkupSafe==3.0.2
mcp==1.8.0
mysql-connector-python==9.3.0
proxy_tools==0.1.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
pythonnet==3.0.5
#pywebview==5.4
sniffio==1.3.1
sse-starlette==2.3.4
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
wincertstore==0.2

# --- LangChain & RAG Dependencies ---

langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
langchain_community==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
tiktoken==0.7.0

# RAG & 数据处理相关
chromadb==1.0.13
langchain-chroma==0.2.4
# sentence-transformers需要torch，使用CPU版本避免下载2GB+的CUDA库
torch==2.7.1 
sentence-transformers==3.0.1
pymupdf==1.24.7
unstructured==0.14.9
langchain-huggingface>=0.1.0

# 数据可视化依赖
matplotlib>=3.5.0
seaborn>=0.12.0

# 因果分析和图处理相关
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.0
causal-learn>=0.1.3.3

openai>=1.55.3

# 数据库迁移工具
alembic==1.17.0
PyMySQL==1.1.2